+++ src/main/java/com/atlauncher/gui/tabs/settings/GeneralSettingsTab.java
@@ -49,9 +49,9 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
     private final JComboBox<String> language;
     private final JComboBox<ComboItem<String>> theme;
     private final JComboBox<ComboItem<String>> dateFormat;
+    private final JComboBox<ComboItem<String>> instanceTitleFormat;
     private final JComboBox<ComboItem<Integer>> selectedTabOnStartup;
     private final JCheckBox sortPacksAlphabetically;
-    private final JCheckBox showPackNameAndVersion;
     private final JCheckBox keepLauncherOpen;
     private final JCheckBox enableConsole;
     private final JCheckBox enableTrayIcon;
@@ -162,6 +162,32 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
 
         add(dateFormat, gbc);
 
+        // Instance Title Format
+
+        gbc.gridx = 0;
+        gbc.gridy++;
+        gbc.insets = UIConstants.LABEL_INSETS;
+        gbc.anchor = GridBagConstraints.BASELINE_TRAILING;
+
+        JLabelWithHover instanceTitleFormatLabel = new JLabelWithHover(GetText.tr("Instance Title Format") + ":",
+                HELP_ICON, GetText.tr("This controls the format that instances titles are shown as."));
+
+        add(instanceTitleFormatLabel, gbc);
+
+        gbc.gridx++;
+        gbc.insets = UIConstants.FIELD_INSETS;
+        gbc.anchor = GridBagConstraints.BASELINE_LEADING;
+        instanceTitleFormat = new JComboBox<>();
+
+        for (String format : Constants.INSTANCE_TITLE_FORMATS) {
+            instanceTitleFormat.addItem(new ComboItem<>(format, String.format(format, GetText.tr("Instance Name"),
+                    GetText.tr("Pack Name"), GetText.tr("Pack Version"), GetText.tr("Minecraft Version"))));
+        }
+
+        instanceTitleFormat.setSelectedItem(App.settings.instanceTitleFormat);
+
+        add(instanceTitleFormat, gbc);
+
         // Selected tab on startup
 
         gbc.gridx = 0;
@@ -219,25 +245,6 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
             sortPacksAlphabetically.setSelected(true);
         }
         add(sortPacksAlphabetically, gbc);
-
-        // Show Pack Name & Version
-
-        gbc.gridx = 0;
-        gbc.gridy++;
-        gbc.insets = UIConstants.LABEL_INSETS;
-        gbc.anchor = GridBagConstraints.BASELINE_TRAILING;
-        JLabelWithHover showPackNameAndVersionLabel = new JLabelWithHover(GetText.tr("Show Pack Name & Version") + "?",
-                HELP_ICON, GetText.tr("If you want to show the packs name and version on your instances."));
-        add(showPackNameAndVersionLabel, gbc);
-
-        gbc.gridx++;
-        gbc.insets = UIConstants.CHECKBOX_FIELD_INSETS;
-        gbc.anchor = GridBagConstraints.BASELINE_LEADING;
-        showPackNameAndVersion = new JCheckBox();
-        if (App.settings.showPackNameAndVersion) {
-            showPackNameAndVersion.setSelected(true);
-        }
-        add(showPackNameAndVersion, gbc);
 
         // Keep Launcher Open
 
@@ -427,7 +434,8 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
     }
 
     public boolean needToReloadInstancesPanel() {
-        return showPackNameAndVersion.isSelected() != App.settings.showPackNameAndVersion;
+        return !(((ComboItem<String>) instanceTitleFormat.getSelectedItem()).getValue())
+                .equals(App.settings.instanceTitleFormat);
     }
 
     public boolean needToReloadLanguage() {
@@ -440,9 +448,9 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
         App.settings.language = (String) language.getSelectedItem();
         App.settings.theme = ((ComboItem<String>) theme.getSelectedItem()).getValue();
         App.settings.dateFormat = ((ComboItem<String>) dateFormat.getSelectedItem()).getValue();
+        App.settings.instanceTitleFormat = ((ComboItem<String>) instanceTitleFormat.getSelectedItem()).getValue();
         App.settings.selectedTabOnStartup = ((ComboItem<Integer>) selectedTabOnStartup.getSelectedItem()).getValue();
         App.settings.sortPacksAlphabetically = sortPacksAlphabetically.isSelected();
-        App.settings.showPackNameAndVersion = showPackNameAndVersion.isSelected();
         App.settings.keepLauncherOpen = keepLauncherOpen.isSelected();
         App.settings.enableConsole = enableConsole.isSelected();
         App.settings.enableTrayMenu = enableTrayIcon.isSelected();