+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### New Features
 - Add support to export CurseForge instances with Fabric
+- Add setting to customise the instance title format
 
 ### Fixes
 - Some packs on CurseForge not being able to be installed due to having bad manifestVersion